LIBSEARCH-79. Added database migrations from NCSU QuickSearch
Ran the railties migration:

```
> rails railties:install:migrations
```

This added a number of files into the "db/migrate" directory.

Edited each of the files, adding "[4.2]" to each of the "ActiveRecord::Migration" superclasses, i.e.:

```
class CreateEvents < ActiveRecord::Migration[4.2]
```

because a version number is now required in Rails 5.1.

Ran "rails db:reset"  and "rails db:migrate" to recreate the databases:

```
> rails db:reset
> rails db:migrate
```

https://issues.umd.edu/browse/LIBSEARCH-79