LIBSEARCH-79. Modified .rubocop.yml config to correctly process files
In Rubocop v0.56.0, the "Include" handler was changed to only include
the listed files, not merge them with the "Include" list in the Rubocop
default.yml file. Since the current "Include" files are already in the
default.yml file, removed "Include" directive.

Added "inherit_mode" as "merge" so if we do eventually have an
"Include" directive, it will merge with the default.yml directives.

This caused a lot more files to be scanned, and the resulting Rubocop
offenses have either been suppressed or corrected.

https://issues.umd.edu/browse/LIBSEARCH-79